+++ src/interface.ts
@@ -1,7 +1,10 @@
-import { QuickPickItem, window } from "vscode";
-import { apktool } from "./tools";
+import * as path from "path";
+import * as fs from "fs";
+import { commands, QuickPickItem, Uri, window } from "vscode";
+import { apktool, initGitDir, jadx } from "./tools";
 import { outputChannel } from "./common";
 import { quickPickUtil } from "./quick-pick.util";
+import { Quark } from "./quark-tools";
 
 export namespace UI {
     /**
@@ -21,13 +24,15 @@ export namespace UI {
             matchOnDescription: true,
             ignoreFocusOut: true,
         });
-        return result ? result.map<string>((item) => item.label) : undefined;
+        return result
+            ? result.map<string>((item) => item.label.split(" ")[0])
+            : undefined;
     }
 
     /**
      * Show a APK file chooser window and decompile that APK.
      */
-    export async function decompileAPK(): Promise<void> {
+    export async function openApkFile(): Promise<void> {
         // browse for an APK file
         const result = await window.showOpenDialog({
             canSelectFolders: false,
@@ -39,16 +44,71 @@ export namespace UI {
         if (result && result.length === 1) {
             const args = await showArgsQuickPick(
                 quickPickUtil.getQuickPickItems("decodeQuickPickItems"),
-                "Additional apktool/jadx arguments"
+                "Additional features & Apktool/Jadx arguments"
             );
             if (args) {
                 const decompileJavaIndex = args.indexOf("decompile_java");
+                const quarkAnalysisIndex = args.indexOf("quark_analysis");
+                const jadxOptionsIndex = args.indexOf("--deobf");
                 let decompileJava = false;
+                let quarkAnalysis = false;
+                let jadxArgs: string[] = [];
+                if (jadxOptionsIndex > -1) {
+                    jadxArgs = args.splice(jadxOptionsIndex, 1);
+                }
                 if (decompileJavaIndex > -1) {
                     decompileJava = true;
                     args.splice(decompileJavaIndex, 1);
                 }
-                await apktool.decodeAPK(result[0].fsPath, args, decompileJava);
+                if (quarkAnalysisIndex > -1) {
+                    quarkAnalysis = true;
+                    args.splice(quarkAnalysisIndex, 1);
+                    if (!Quark.checkQuarkInstalled()) {
+                        quarkAnalysis = false;
+                        window.showErrorMessage(
+                            "APKLab: Quark command not found, \
+                            please make sure you have installed python3 and Quark-Engine. \
+                            Check here to install Quark-Engine: \
+                            https://github.com/quark-engine/quark-engine"
+                        );
+                        return;
+                    }
+                }
+
+                // project directory name
+                const apkFilePath = result[0].fsPath;
+                let projectDir = path.join(
+                    path.dirname(apkFilePath),
+                    path.parse(apkFilePath).name
+                );
+                // don't delete the existing dir if it already exists
+                while (fs.existsSync(projectDir)) {
+                    projectDir = projectDir + "1";
+                }
+
+                // decode APK
+                await apktool.decodeAPK(apkFilePath, projectDir, args);
+
+                // decompile APK
+                if (decompileJava) {
+                    await jadx.decompileAPK(apkFilePath, projectDir, jadxArgs);
+                }
+                // quark analysis
+                if (quarkAnalysis) {
+                    await Quark.analyzeAPK(apkFilePath, projectDir);
+                }
+
+                // Initialize project dir as git repo
+                await initGitDir(projectDir, "Initial APKLab project");
+
+                // open project dir in a new window
+                if (!process.env["TEST"]) {
+                    await commands.executeCommand(
+                        "vscode.openFolder",
+                        Uri.file(projectDir),
+                        true
+                    );
+                }
             }
         } else {
             outputChannel.appendLine("APKLAB: no APK file was chosen");
@@ -62,7 +122,7 @@ export namespace UI {
     export async function rebuildAPK(apktoolYmlPath: string): Promise<void> {
         const args = await showArgsQuickPick(
             quickPickUtil.getQuickPickItems("rebuildQuickPickItems"),
-            "Additional apktool arguments"
+            "Additional Apktool arguments"
         );
         if (args) {
             await apktool.rebuildAPK(apktoolYmlPath, args);